add basic sample method
