+++ lib/filter.js
@@ -1,6 +1,6 @@
 const R = require('ramda');
 const jBinary = require('jbinary');
-const binaryTypeset = require('./binaryTypeset');
+const { mapBinary } = require('./helpers');
 
 const between = R.curry((val, low, high) =>
   R.and(R.gt(val, low), R.lt(val, high))
@@ -57,33 +57,9 @@ const filterPoints = (options = {}, points = []) => {
   return points.filter(filterWithOptions);
 };
 
-const filterBinary = R.curry((options, binary) => {
-  const data = binary.readAll();
-  const header = data.header;
-  const pointFormat = binary.typeSet[`pointFormat${header.pointFormat}`];
-  const numberOfPoints =
-    header.versionMinor < 4
-      ? header.legacyNumberOfPoints
-      : header.numberOfPoints;
-  const points = binary.read(
-    ['array', pointFormat, numberOfPoints],
-    header.offsetToPoints
-  );
-  const oldLength = data.points.length;
-  data.points = filterPoints(options, points);
-
-  // Must create new instance rather than mutate existing because memory is statically allocated.
-  // New size is based on new number of points
-  // https://github.com/jDataView/jBinary/issues/48
-  const newLength =
-    binary.view.byteLength -
-    (oldLength - data.points.length) *
-      binary.typeSet.sizes[`pointFormat${header.pointFormat}`];
-  const newBinary = new jBinary(newLength, binaryTypeset);
-  newBinary.writeAll(data);
-
-  return newBinary;
-});
+const filterBinary = R.curry((options, binary) =>
+  mapBinary(filterPoints, options, binary)
+);
 
 module.exports = {
   filterBinary: filterBinary,
